GH-10: Refactoring.
