+++ tests/test_text_classifier_trainer.py
@@ -1,24 +1,20 @@
 import shutil
 
 from flair.data_fetcher import NLPTaskDataFetcher, NLPTask
-from flair.embeddings import CharLMEmbeddings, WordEmbeddings
+from flair.embeddings import WordEmbeddings
 from flair.models.text_classification_model import TextClassifier
 from flair.trainers.text_classification_trainer import TextClassifierTrainer
 
 
 def test_training():
     corpus = NLPTaskDataFetcher.fetch_data(NLPTask.IMDB)
-
-    charlm_embedding_forward = CharLMEmbeddings('news-forward')
-    charlm_embedding_backward = CharLMEmbeddings('news-backward')
-    glove_embedding = WordEmbeddings('en-glove')
-
     label_dict = corpus.make_label_dictionary()
 
-    model = TextClassifier([charlm_embedding_forward, charlm_embedding_backward, glove_embedding], 128, 1, False, False, label_dict, False)
+    glove_embedding = WordEmbeddings('en-glove')
+    model = TextClassifier([glove_embedding], 128, 1, False, False, label_dict, False)
 
     trainer = TextClassifierTrainer(model, corpus, label_dict, False)
+    trainer.train('./results', max_epochs=2)
 
-    trainer.train('./results', max_epochs=5)
-
+    # clean up results directory
     shutil.rmtree('./results')
